requirements.txt
Django~=2.2.9
djangorestframework
django-filter
python-dateutil
django-rank-query
